Create SPARQL_query
Note: This query is not perfect or streamlined for every collection; modifications are needed depending on how much metadata is included in the FranceArchives portal, for example whether subjects are specified. It is designed to get the collection and all sub-components, the top-level identifier, and the labels for all rico:hasOrHadSubject tags (<subject> and <geogname> for the corresponding EAD tags).